Move WrappedSystem into transaction_executor
